textSummarize/pyproject.toml
[project]
name = "textsummarize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
]
